+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/ProgressiveMediaSource.java
@@ -22,6 +22,7 @@ import android.os.Looper;
 import androidx.annotation.GuardedBy;
 import androidx.annotation.Nullable;
 import androidx.media3.common.C;
+import androidx.media3.common.Format;
 import androidx.media3.common.MediaItem;
 import androidx.media3.common.Timeline;
 import androidx.media3.common.util.Consumer;
@@ -32,13 +33,16 @@ import androidx.media3.datasource.TransferListener;
 import androidx.media3.exoplayer.drm.DefaultDrmSessionManagerProvider;
 import androidx.media3.exoplayer.drm.DrmSessionManager;
 import androidx.media3.exoplayer.drm.DrmSessionManagerProvider;
+import androidx.media3.exoplayer.trackselection.ExoTrackSelection;
 import androidx.media3.exoplayer.upstream.Allocator;
 import androidx.media3.exoplayer.upstream.DefaultLoadErrorHandlingPolicy;
 import androidx.media3.exoplayer.upstream.LoadErrorHandlingPolicy;
 import androidx.media3.exoplayer.util.ReleasableExecutor;
 import androidx.media3.extractor.DefaultExtractorsFactory;
 import androidx.media3.extractor.Extractor;
+import androidx.media3.extractor.ExtractorOutput;
 import androidx.media3.extractor.ExtractorsFactory;
+import androidx.media3.extractor.TrackOutput;
 import com.google.common.base.Supplier;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.util.concurrent.Executor;
@@ -69,7 +73,8 @@ public final class ProgressiveMediaSource extends BaseMediaSource
     private LoadErrorHandlingPolicy loadErrorHandlingPolicy;
     private int continueLoadingCheckIntervalBytes;
     @Nullable private Supplier<ReleasableExecutor> downloadExecutorSupplier;
-    private boolean suppressPrepareError;
+    private int singleTrackId;
+    @Nullable private Format singleTrackFormat;
 
     /**
      * Creates a new factory for {@link ProgressiveMediaSource}s.
@@ -191,17 +196,21 @@ public final class ProgressiveMediaSource extends BaseMediaSource
     }
 
     /**
-     * Allow {@link MediaPeriod} preparation to {@linkplain
-     * MediaPeriod.Callback#onPrepared(MediaPeriod) complete} despite an error that would have
-     * otherwise blocked it.
+     * Allows the {@link ProgressiveMediaSource} to complete preparation without reading any data.
      *
-     * <p>If set to true, an error that would normally be thrown from {@link
-     * MediaPeriod#maybeThrowPrepareError()} (e.g. a {@link DataSource#open} error like HTTP 404) is
-     * instead suppressed and preparation is completed with no tracks.
+     * <p>This must only be set if the source is guaranteed to emit a single track with the provided
+     * ID and format.
+     *
+     * <p>Data will only be loaded if the track is selected with {@link
+     * MediaPeriod#selectTracks(ExoTrackSelection[], boolean[], SampleStream[], boolean[], long)}
+     *
+     * @param trackId The ID of the track to pass to {@link ExtractorOutput#track}
+     * @param format The format of the track to pass to {@link TrackOutput#format}.
      */
     @CanIgnoreReturnValue
-    /* package */ Factory setSuppressPrepareError(boolean suppressPrepareError) {
-      this.suppressPrepareError = suppressPrepareError;
+    /* package */ Factory enableLazyLoadingWithSingleTrack(int trackId, Format format) {
+      this.singleTrackId = trackId;
+      this.singleTrackFormat = checkNotNull(format);
       return this;
     }
 
@@ -252,7 +261,8 @@ public final class ProgressiveMediaSource extends BaseMediaSource
           drmSessionManagerProvider.get(mediaItem),
           loadErrorHandlingPolicy,
           continueLoadingCheckIntervalBytes,
-          suppressPrepareError,
+          singleTrackId,
+          singleTrackFormat,
           downloadExecutorSupplier);
     }
 
@@ -273,7 +283,19 @@ public final class ProgressiveMediaSource extends BaseMediaSource
   private final DrmSessionManager drmSessionManager;
   private final LoadErrorHandlingPolicy loadableLoadErrorHandlingPolicy;
   private final int continueLoadingCheckIntervalBytes;
-  private final boolean suppressPrepareError;
+
+  /**
+   * The ID passed to {@link Factory#enableLazyLoadingWithSingleTrack(int, Format)}. Only valid if
+   * {@link #singleTrackFormat} is non-null.
+   */
+  private final int singleTrackId;
+
+  /**
+   * The {@link Format} passed to {@link Factory#enableLazyLoadingWithSingleTrack(int, Format)}, or
+   * {@code null} if not set.
+   */
+  @Nullable private final Format singleTrackFormat;
+
   @Nullable private final Supplier<ReleasableExecutor> downloadExecutorSupplier;
 
   private boolean timelineIsPlaceholder;
@@ -292,7 +314,8 @@ public final class ProgressiveMediaSource extends BaseMediaSource
       DrmSessionManager drmSessionManager,
       LoadErrorHandlingPolicy loadableLoadErrorHandlingPolicy,
       int continueLoadingCheckIntervalBytes,
-      boolean suppressPrepareError,
+      int singleTrackId,
+      @Nullable Format singleTrackFormat,
       @Nullable Supplier<ReleasableExecutor> downloadExecutorSupplier) {
     this.mediaItem = mediaItem;
     this.dataSourceFactory = dataSourceFactory;
@@ -300,7 +323,8 @@ public final class ProgressiveMediaSource extends BaseMediaSource
     this.drmSessionManager = drmSessionManager;
     this.loadableLoadErrorHandlingPolicy = loadableLoadErrorHandlingPolicy;
     this.continueLoadingCheckIntervalBytes = continueLoadingCheckIntervalBytes;
-    this.suppressPrepareError = suppressPrepareError;
+    this.singleTrackFormat = singleTrackFormat;
+    this.singleTrackId = singleTrackId;
     this.timelineIsPlaceholder = true;
     this.timelineDurationUs = C.TIME_UNSET;
     this.downloadExecutorSupplier = downloadExecutorSupplier;
@@ -359,7 +383,8 @@ public final class ProgressiveMediaSource extends BaseMediaSource
         allocator,
         localConfiguration.customCacheKey,
         continueLoadingCheckIntervalBytes,
-        suppressPrepareError,
+        singleTrackId,
+        singleTrackFormat,
         Util.msToUs(localConfiguration.imageDurationMs),
         downloadExecutorSupplier != null ? downloadExecutorSupplier.get() : null);
   }